add 'mute voice' for convenience
